+++ apps/frontend-pwa/src/pages/course/[courseId]/repetition.tsx
@@ -1,5 +1,0 @@
-function Repetition() {
-  return <div>hello world</div>
-}
-
-export default Repetition