+++ CHANGELOG.md
@@ -2,6 +2,25 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [2.0.0-alpha.102](https://github.com/uzh-bf/klicker-uzh/compare/v2.0.0-alpha.101...v2.0.0-alpha.102) (2022-11-11)
+
+
+### Bug Fixes
+
+* ensure that courseId is fetched for learning element repetition ([8df5a12](https://github.com/uzh-bf/klicker-uzh/commit/8df5a12084531988f114500afc05b96b558f2c50))
+
+## [2.0.0-alpha.101](https://github.com/uzh-bf/klicker-uzh/compare/v2.0.0-alpha.100...v2.0.0-alpha.101) (2022-11-10)
+
+
+### Features
+
+* learning element repetition on index page ([8575bc9](https://github.com/uzh-bf/klicker-uzh/commit/8575bc944eb8346d5d6db5fd0cfbf32086b84519))
+
+
+### Other
+
+* **packages/prisma:** update seed with draft microlearning bf1 week 09 ([7850490](https://github.com/uzh-bf/klicker-uzh/commit/7850490d9a0c219d75f8c73393f1c9e3eda36e4d))
+
 ## [2.0.0-alpha.100](https://github.com/uzh-bf/klicker-uzh/compare/v2.0.0-alpha.99...v2.0.0-alpha.100) (2022-11-10)